+++ app/javascript/react/components/MetadataEntry/FunderForm.jsx
@@ -1,9 +1,9 @@
-import React, {useRef, useState} from 'react';
+import React, {useEffect, useRef, useState} from 'react';
 // see https://formik.org/docs/tutorial for basic tutorial, yup is easy default for validation w/ formik
 import {Field, Form, Formik} from 'formik';
 import axios from 'axios';
 import PropTypes from 'prop-types';
-import FunderAutocomplete from './FunderAutocomplete';
+import RorAutocomplete from './RorAutocomplete';
 import {showModalYNDialog, showSavedMsg, showSavingMsg} from '../../../lib/utils';
 
 function FunderForm({
@@ -14,6 +14,15 @@ function FunderForm({
   // the follow autocomplete items are lifted up state that is normally just part of the form, but doesn't work with Formik
   const [acText, setAcText] = useState(contributor.contributor_name || '');
   const [acID, setAcID] = useState(contributor.name_identifier_id || '');
+  const [showSelect, setShowSelect] = useState(null);
+
+  const subSelect = (e) => {
+    const select = e.target;
+    setAcID(select.value);
+    setAcText(select.selectedOptions[0].text);
+    formRef.current.handleSubmit();
+    setShowSelect(null);
+  };
 
   const submitForm = (values) => {
     showSavingMsg();
@@ -26,7 +35,8 @@ function FunderForm({
         id: values.id,
         contributor_name: acText,
         contributor_type: 'funder',
-        identifier_type: 'crossref_funder_id', // needs to be set for datacite mapping, even if no id gotten from crossref
+        // needs to be set for datacite mapping, even if no id gotten from crossref
+        identifier_type: acID.includes('ror.org') ? 'ror' : 'crossref_funder_id',
         name_identifier_id: acID,
         award_number: values.award_number,
         award_description: values.award_description,
@@ -55,6 +65,15 @@ function FunderForm({
     });
   };
 
+  useEffect(() => {
+    const group = groupings?.find((g) => g.name_identifier_id === acID);
+    if (group) {
+      setShowSelect(group);
+    } else {
+      setShowSelect(null);
+    }
+  }, [acID]);
+
   return (
     <Formik
       initialValues={
@@ -73,13 +92,12 @@ function FunderForm({
         <Form className="c-input__inline">
           <Field name="id" type="hidden" />
           <div className="c-input">
-            <FunderAutocomplete
+            <RorAutocomplete
               formRef={formRef}
               acText={acText}
               setAcText={setAcText}
               acID={acID}
               setAcID={setAcID}
-              groupings={groupings}
               controlOptions={
                 {
                   htmlId: `contrib_${contributor.id}`,
@@ -88,6 +106,15 @@ function FunderForm({
                 }
               }
             />
+            {showSelect && (
+              <>
+                <label htmlFor="subfunder_select" className="c-input__label" style={{marginTop: '1em'}}>{showSelect.group_label}</label>
+                <select id="subfunder_select" className="c-input__select" onChange={subSelect}>
+                  <option value="">- Select one -</option>
+                  {showSelect.json_contains.map((i) => <option key={i.name_identifier_id} value={i.name_identifier_id}>{i.contributor_name}</option>)}
+                </select>
+              </>
+            )}
           </div>
           <div className="c-input">
             <label className="c-input__label" htmlFor={`contributor_award_number__${contributor.id}`}>Award
